requirements.txt
flask
youtube-transcript-api
nltk
reportlab
mistralai
markdown
python-dotenv